backend/pyproject.toml
[project]
name = "simple-dictionary"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.3",
    "django-cors-headers>=4.9.0",
    "django-filter>=25.2",
    "django-rest-swagger>=2.2.0",
    "djangorestframework>=3.17.1",
    "djangorestframework-simplejwt>=5.5.1",
]
